# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: ContaFácil API — gestão contábil
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /competencias:
    get:
      operationId: listarCompetencias
      tags: [competencias]
      summary: Lista competências com resumo
      responses:
        "200":
          description: Lista de competências
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CompetenciaComResumo"
    post:
      operationId: abrirCompetencia
      tags: [competencias]
      summary: Abre uma nova competência (gera checklist e pagamentos)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AbrirCompetenciaInput"
      responses:
        "200":
          description: Competência criada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Competencia"
  /competencias/{id}:
    get:
      operationId: getCompetencia
      tags: [competencias]
      summary: Obtém uma competência com resumo
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            minimum: 1
      responses:
        "200":
          description: Competência
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompetenciaComResumo"
        "404":
          $ref: "#/components/responses/Erro"
    delete:
      operationId: removerCompetencia
      tags: [competencias]
      summary: Remove competência
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            minimum: 1
      responses:
        "204":
          description: Removida
  /competencias/{id}/sincronizar:
    post:
      operationId: sincronizarCompetencia
      tags: [competencias]
      summary: Acrescenta ao mês os clientes e obrigações incluídos depois da abertura
      description: Idempotente — só cria o que falta; itens e pagamentos existentes não são tocados.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            minimum: 1
      responses:
        "200":
          description: O que foi acrescentado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResultadoSincronizacao"
        "404":
          $ref: "#/components/responses/Erro"
  /competencias/{id}/checklist:
    get:
      operationId: listarChecklist
      tags: [checklist]
      summary: Lista itens do checklist da competência
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            minimum: 1
      responses:
        "200":
          description: Itens do checklist
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChecklistItem"
  /competencias/{id}/pagamentos:
    get:
      operationId: listarPagamentos
      tags: [pagamentos]
      summary: Lista pagamentos da competência
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            minimum: 1
      responses:
        "200":
          description: Pagamentos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Pagamento"
...
  /jobs:
    get:
      operationId: listarJobs
      tags: [jobs]
      summary: Últimas tarefas da fila (só admin)
      parameters:
        - name: status
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/StatusJob"
      responses:
        "200":
          description: Lista de tarefas
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Job"
  /jobs/executar:
    post:
      operationId: executarJobs
      tags: [jobs]
      summary: Processa a fila por até alguns minutos (gatilho do agendador externo)
      security:
        - tokenJobs: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExecutarJobsInput"
      responses:
        "200":
          description: Resultado da rodada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResultadoJobs"
        "401":
          $ref: "#/components/responses/Erro"
components:
  securitySchemes:
    tokenJobs:
      type: http
      scheme: bearer
      description: "`Authorization: Bearer ${TOKEN_JOBS}`"
  responses:
    Erro:
      description: Erro tratado pela API
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Erro"
  schemas:
...
    Erro:
      type: object
      required: [error, codigo]
      properties:
        error: { type: string }
        codigo: { type: string }
        details: {}
        issues:
          type: array
          items: { type: object, additionalProperties: true }
...
    StatusJob:
      type: string
      enum: [pendente, executando, concluido, falhou]
    Job:
      type: object
      required: [id, tipo, chave, status, tentativas, maxTentativas, executarEm, iniciadoEm, concluidoEm, erro, criadoEm]
      properties:
        id: { type: integer }
        contaId: { type: ["integer", "null"] }
        tipo: { type: string }
        chave: { type: ["string", "null"] }
        dados: { type: object, additionalProperties: true }
        status: { $ref: "#/components/schemas/StatusJob" }
        tentativas: { type: integer }
        maxTentativas: { type: integer }
        executarEm: { type: string, format: date-time }
        iniciadoEm: { type: ["string", "null"], format: date-time }
        concluidoEm: { type: ["string", "null"], format: date-time }
        erro: { type: ["string", "null"] }
        criadoEm: { type: string, format: date-time }
    ExecutarJobsInput:
      type: object
      properties:
        limiteMs: { type: integer, minimum: 1000, maximum: 300000 }
        # Tipos recorrentes para rodar agora, fora da chave diária (ex.: depois
        # de corrigir um cadastro): `vencimentos-d3`, `honorarios-vencidos`, `limpeza`.
        agora:
          type: array
          maxItems: 10
          items: { type: string, minLength: 1, maxLength: 60 }
    ResultadoJobs:
      type: object
      required: [executados, falhas, restantes]
      properties:
        executados: { type: integer }
        falhas: { type: integer }
        restantes: { type: integer }
...
    Resumo:
      type: object
      required: [obrigacoes, pagamentos]
      properties:
        obrigacoes:
          type: object
          required: [total, feitos, emitidos, pendentes]
          properties:
            total: { type: integer }
            # `feitos` = enviados ao cliente, o fim do ciclo.
            feitos: { type: integer }
            emitidos: { type: integer }
            pendentes: { type: integer }
        pagamentos:
          type: object
          required: [total, pagos, pendentes, recebido, aReceber]
          properties:
            total: { type: integer }
            pagos: { type: integer }
            pendentes: { type: integer }
            recebido: { type: string, pattern: '^-?\d+(\.\d{1,2})?$' }
            aReceber: { type: string, pattern: '^-?\d+(\.\d{1,2})?$' }
...
    Competencia:
      type: object
      required: [id, ano, mes]
      properties:
        id: { type: integer }
        ano: { type: integer, minimum: 2000, maximum: 2100 }
        mes: { type: integer, minimum: 1, maximum: 12 }
        rotulo: { type: ["string", "null"] }
    CompetenciaComResumo:
      type: object
      required: [id, ano, mes, resumo]
      properties:
        id: { type: integer }
        ano: { type: integer, minimum: 2000, maximum: 2100 }
        mes: { type: integer, minimum: 1, maximum: 12 }
        rotulo: { type: ["string", "null"] }
        resumo:
          $ref: "#/components/schemas/Resumo"
    AbrirCompetenciaInput:
      type: object
      required: [ano, mes]
      properties:
        ano: { type: integer, minimum: 2000, maximum: 2100 }
        mes: { type: integer, minimum: 1, maximum: 12 }
        # Para meses anteriores ao início do uso do sistema: gera só os
        # honorários (pagamentos), sem o checklist de obrigações daquele mês.
        somenteHonorarios: { type: boolean }
    ResultadoSincronizacao:
      type: object
      required: [itensCriados, pagamentosCriados]
      properties:
        itensCriados: { type: integer }
        pagamentosCriados: { type: integer }
    ChecklistItem:
      type: object
      required: [id, status, clienteId, cliente, tipoObrigacaoId, obrigacao, ordem, anexos, enviadoEm, visualizadoEm]
      properties:
        id: { type: integer }
        status:
          type: string
          enum: [pendente, emitido, enviado, nao_aplica]
        clienteId: { type: integer }
        codigo: { type: ["integer", "null"], minimum: 0 }
        cliente: { type: string }
        vencimento: { type: ["string", "null"], format: date }
        tipoObrigacaoId: { type: integer }
        obrigacao: { type: string }
        ordem: { type: integer, minimum: 0 }
        # Guias anexadas (confirmadas) e o último envio com protocolo.
        anexos: { type: integer }
        enviadoEm: { type: ["string", "null"], format: date-time }
        visualizadoEm: { type: ["string", "null"], format: date-time }
...
    Pagamento:
      type: object
      required: [id, status, clienteId, cliente]
      properties:
        id: { type: integer }
        status:
          type: string
          enum: [pendente, pago, isento]
        valor: { type: ["string", "null"], pattern: '^-?\d+(\.\d{1,2})?$' }
        dataPagamento: { type: ["string", "null"], format: date }
        vencimento: { type: ["string", "null"], format: date }
        forma: { type: ["string", "null"] }
        observacao: { type: ["string", "null"] }
        clienteId: { type: integer }
        codigo: { type: ["integer", "null"], minimum: 0 }
        cliente: { type: string }
...